Handle inconsistent state when destroying windows
window.image_acquired_semaphores[i] contains uninitialized values.

This means that if VulkanContext::_update_swap_chain had failed early
(this can happen for valid reasons and due to unexpected failures) then
_clean_up_swap_chain will try to destroy an invalid handle.
This fix always sets the handle to VK_NULL_HANDLE.

Likewise perform the checks in various other resources in
_clean_up_swap_chain in case the resources were partially initialized
and are now in an inconsistent state. The API should guarantee the call
to be valid if the handle was VK_NULL_HANDLE, but I don't trust driver
implementations to handle that case.

Affects #81670